+++ src/app/(with-menu)/layout.tsx
@@ -1,0 +1,9 @@
+import { LayoutMenu } from '@/layouts/Menu';
+
+export default function WithMenuLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <LayoutMenu>{children}</LayoutMenu>;
+}
