+++ src/app/(with-menu)/menu-sample/menu-3/page.tsx
@@ -1,13 +1,8 @@
 'use client';
 
-import { useEffect } from 'react';
 import { Index } from './components';
 
 export default function AccordionSample() {
-  useEffect(() => {
-    import('bootstrap/dist/js/bootstrap');
-  }, []);
-
   return (
     <>
       <Index />
