+++ src/app/(without-menu)/accordion-sample/page.tsx
@@ -1,6 +1,13 @@
+import { BaseButton } from '@/components/elements/buttons/BaseButton';
 import { Accordion } from '@/components/elements/accordion';
+import { LayoutDefault } from '@/layouts/Default';
 
-export const Index = () => {
+export const metadata = {
+  title: 'アコーディオンサンプル | nextjs-bootstrap',
+  description: '',
+};
+
+export default function AccordionSample() {
   return (
     <>
       <h2 className="h3 pt-3 pb-2 mb-3 border-bottom">
@@ -35,6 +42,13 @@ export const Index = () => {
           </>
         }
       ></Accordion>
+      <div className="position-absolute bottom-0 end-0 m-5">
+        <BaseButton
+          url="/"
+          className="btn-primary btn-sm"
+          title="トップに戻る"
+        />
+      </div>
     </>
   );
-};
+}
