+++ src/app/(without-menu)/layout.tsx
@@ -1,11 +1,6 @@
 import { LayoutDefault } from '@/layouts/Default';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
-export const metadata = {
-  title: 'フォームサンプル | nextjs-bootstrap',
-  description: '',
-};
-
 export default function AccordionSample({
   children,
 }: {
