--- src/app/menu-sample/page.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import { useEffect } from 'react';
-
-export default function LayoutSample() {
-  useEffect(() => {
-    import('bootstrap/dist/js/bootstrap');
-  }, []);
-
-  return <></>;
-}
