+++ src/components/layout/LayoutMenuHeader/index.tsx
@@ -5,9 +5,9 @@ import { NavBarToggleButton } from '@/components/elements/NavbarToggleButton';
 export const LayoutSideMenuHeader = () => {
   return (
     <header className="navbar navbar-dark sticky-top bg-dark flex-md-nowrap p-0 shadow">
-      <a className="navbar-brand col-md-3 col-lg-2 me-0 px-3" href="./">
+      <Link className="navbar-brand col-md-3 col-lg-2 me-0 px-3" href="/">
         ロゴ
-      </a>
+      </Link>
       <NavBarToggleButton
         targetId="sidebarMenu"
         className="position-absolute d-md-none collapsed"
